sdk/sqreen-instrumentation-tool: move tool validation tests into the tool dir
