+++ sdk/sqreen-instrumentation-tool/testdata/hello-example/output.txt
