+++ sdk/sqreen-instrumentation-tool/tool_test.go
@@ -1,8 +1,8 @@
-// Copyright (c) 2016 - 2019 Sqreen. All Rights Reserved.
+// Copyright (c) 2016 - 2020 Sqreen. All Rights Reserved.
 // Please refer to our terms for more information:
 // https://www.sqreen.io/terms.html
 
-package sqreen_test
+package main_test
 
 import (
 	"fmt"
